# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Transfert Bancaire",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api-transfert.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/oauth/token": {
            "post": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Obtenir un token d'accès",
                "description": "Authentification OAuth2 pour obtenir un access token. Route générée par Passport.",
                "operationId": "26570df72160b2d15db22bb80394e1f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grant_type",
                                    "client_id",
                                    "client_secret",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "grant_type": {
                                        "type": "string",
                                        "enum": [
                                            "password"
                                        ],
                                        "example": "password"
                                    },
                                    "client_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "client_secret": {
                                        "type": "string",
                                        "example": "client-secret-here"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "scope": {
                                        "type": "string",
                                        "example": "*"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token obtenu avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 31536000
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur d'authentification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "invalid_credentials"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The user credentials were incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/oauth/token/refresh": {
            "post": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Rafraîchir un token d'accès",
                "description": "Utiliser un refresh token pour obtenir un nouveau access token. Route générée par Passport.",
                "operationId": "a79c3077c57526b719524f54f512417f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grant_type",
                                    "refresh_token",
                                    "client_id",
                                    "client_secret"
                                ],
                                "properties": {
                                    "grant_type": {
                                        "type": "string",
                                        "enum": [
                                            "refresh_token"
                                        ],
                                        "example": "refresh_token"
                                    },
                                    "refresh_token": {
                                        "type": "string",
                                        "example": "refresh-token-here"
                                    },
                                    "client_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "client_secret": {
                                        "type": "string",
                                        "example": "client-secret-here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouveau token obtenu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 31536000
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/oauth/clients": {
            "get": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Lister les clients OAuth2",
                "description": "Récupérer la liste des clients OAuth2 enregistrés. Route générée par Passport.",
                "operationId": "2757e410076654b7592609f6f52bceaa",
                "responses": {
                    "200": {
                        "description": "Liste des clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "secret": {
                                                "type": "string"
                                            },
                                            "redirect": {
                                                "type": "string"
                                            },
                                            "personal_access_client": {
                                                "type": "boolean"
                                            },
                                            "password_client": {
                                                "type": "boolean"
                                            },
                                            "revoked": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Créer un client OAuth2",
                "description": "Créer un nouveau client OAuth2. Route générée par Passport.",
                "operationId": "c1d3a11633c885f710e9a07462f6041c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "redirect"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My App"
                                    },
                                    "redirect": {
                                        "type": "string",
                                        "example": "http://localhost/callback"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "secret": {
                                            "type": "string"
                                        },
                                        "redirect": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/oauth/clients/{client_id}": {
            "put": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Modifier un client OAuth2",
                "description": "Modifier les informations d'un client OAuth2. Route générée par Passport.",
                "operationId": "f1cf3d2066bb4e4d3c225f7109a195f7",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "redirect"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated App Name"
                                    },
                                    "redirect": {
                                        "type": "string",
                                        "example": "http://localhost/new-callback"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client modifié"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Supprimer un client OAuth2",
                "description": "Supprimer un client OAuth2. Route générée par Passport.",
                "operationId": "20309bdd897c31ea5efcbf8b37352352",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client supprimé"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/oauth/scopes": {
            "get": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Lister les scopes OAuth2",
                "description": "Récupérer la liste des scopes OAuth2 disponibles. Route générée par Passport.",
                "operationId": "8fb2ff58e5deb2170711e1902d035654",
                "responses": {
                    "200": {
                        "description": "Liste des scopes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/oauth/user": {
            "get": {
                "tags": [
                    "OAuth2"
                ],
                "summary": "Informations utilisateur OAuth2",
                "description": "Récupérer les informations de l'utilisateur OAuth2. Route générée par Passport.",
                "operationId": "38d19b733e0a8744955193b6418dedc6",
                "responses": {
                    "200": {
                        "description": "Informations utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string",
                                            "enum": [
                                                "admin",
                                                "client"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
...
    "security": [
        {
            "passport": []
        }
    ]
}